Factor CREATE_HAS_STATIC_MEMBER_VARIABLE into its own file
