+++ tests/Unit/Utilities/Test_TypeTraits.cpp
@@ -417,26 +417,4 @@ static_assert(not is_foo_callable_r_v<int, bar, int, double>,
 static_assert(not is_foobar_callable_v<bar, int, double>,
               "Failed testing CREATE_IS_CALLABLE");
 /// [CREATE_IS_CALLABLE_EXAMPLE]
-
-/// [CREATE_HAS_EXAMPLE]
-CREATE_HAS_STATIC_MEMBER_VARIABLE(foo)
-CREATE_HAS_STATIC_MEMBER_VARIABLE_V(foo)
-CREATE_HAS_STATIC_MEMBER_VARIABLE(foobar)
-CREATE_HAS_STATIC_MEMBER_VARIABLE_V(foobar)
-struct testing_create_has_static_member_variable {
-  static constexpr size_t foo = 1;
-};
-
-static_assert(has_foo_v<testing_create_has_static_member_variable>,
-              "Failed testing CREATE_HAS_STATIC_MEMBER_VARIABLE");
-static_assert(has_foo_v<testing_create_has_static_member_variable, size_t>,
-              "Failed testing CREATE_HAS_STATIC_MEMBER_VARIABLE");
-static_assert(not has_foo_v<testing_create_has_static_member_variable, int>,
-              "Failed testing CREATE_HAS_STATIC_MEMBER_VARIABLE");
-static_assert(not has_foobar_v<testing_create_has_static_member_variable>,
-              "Failed testing CREATE_HAS_STATIC_MEMBER_VARIABLE");
-static_assert(
-    not has_foobar_v<testing_create_has_static_member_variable, size_t>,
-    "Failed testing CREATE_HAS_STATIC_MEMBER_VARIABLE");
-/// [CREATE_HAS_EXAMPLE]
 }  // namespace
